+++ src/onegov.ballot/onegov/ballot/upgrade.py
@@ -119,3 +119,24 @@ def add_municipality_domain(context):
         )
 
     tmp_type.drop(context.operations.get_bind(), checkfirst=False)
+
+
+@upgrade_task('Add party resuts columns')
+def add_party_results_columns(context):
+    if not context.has_column('party_results', 'color'):
+        context.operations.add_column(
+            'party_results',
+            Column('color', Text())
+        )
+
+    if not context.has_column('party_results', 'year'):
+        context.operations.add_column(
+            'party_results',
+            Column('year', Integer())
+        )
+
+    if not context.has_column('party_results', 'total_votes'):
+        context.operations.add_column(
+            'party_results',
+            Column('total_votes', Integer())
+        )